+++ iblrig/version_management.py
@@ -1,6 +1,7 @@
 import logging
 import re
 from collections.abc import Callable
+from functools import cache
 from pathlib import Path
 from subprocess import STDOUT, CalledProcessError, SubprocessError, check_call, check_output
 from typing import Any, Literal
@@ -10,7 +11,7 @@ from packaging import version
 
 from iblrig import __version__
 from iblrig.constants import BASE_DIR, IS_GIT, IS_VENV
-from iblrig.tools import cached_check_output, internet_available, static_vars
+from iblrig.tools import cached_check_output, internet_available
 
 log = logging.getLogger(__name__)
 
@@ -209,7 +210,6 @@ def get_commit_hash(short: bool = True):
     return call_git(*args, on_error='log')
 
 
-@static_vars(is_fetched_already=False)
 def get_remote_tags() -> None:
     """
     Fetch remote Git tags if not already fetched.
@@ -226,20 +226,14 @@ def get_remote_tags() -> None:
     -----
     This method will only work with installations managed through Git.
     """
-    if get_remote_tags.is_fetched_already or not internet_available():
+    if not internet_available():
         return
-    if not IS_GIT:
-        log.error('This installation of iblrig is not managed through git')
     if (branch := get_branch()) is None:
         return
-    try:
-        check_call(['git', 'fetch', 'origin', branch, '-t', '-q', '-f'], cwd=BASE_DIR, timeout=5)
-    except (SubprocessError, CalledProcessError):
-        return
-    get_remote_tags.is_fetched_already = True
+    call_git('fetch', 'origin', branch, '-t', '-q', '-f', on_error='log')
 
 
-@static_vars(changelog=None)
+@cache
 def get_changelog() -> str:
     """
     Retrieve the changelog for the iblrig installation.
@@ -259,20 +253,18 @@ def get_changelog() -> str:
     This method relies on the presence of a CHANGELOG.md file either in the
     repository or locally.
     """
-    if get_changelog.changelog is not None:
-        return get_changelog.changelog
     try:
+        if (branch := get_branch()) is None:
+            raise RuntimeError()
         changelog = requests.get(
-            f'https://raw.githubusercontent.com/int-brain-lab/iblrig/{get_branch()}/CHANGELOG.md', allow_redirects=True
+            f'https://raw.githubusercontent.com/int-brain-lab/iblrig/{branch}/CHANGELOG.md', allow_redirects=True
         ).text
-    except requests.RequestException:
+    except (requests.RequestException, RuntimeError):
         with open(Path(BASE_DIR).joinpath('CHANGELOG.md')) as f:
             changelog = f.read()
-    get_changelog.changelog = changelog
-    return get_changelog.changelog
+    return changelog
 
 
-@static_vars(remote_version=None)
 def get_remote_version() -> version.Version | None:
     """
     Retrieve the remote version of iblrig from the Git repository.
@@ -290,31 +282,11 @@ def get_remote_version() -> version.Version | None:
     -----
     This method will only work with installations managed through Git.
     """
-    if get_remote_version.remote_version is not None:
-        log.debug(f'Using cached remote version: {get_remote_version.remote_version}')
-        return get_remote_version.remote_version
-
-    if not IS_GIT:
-        log.error('Cannot obtain remote version: This installation of iblrig is not managed through git')
-        return None
-
     if not internet_available():
         log.error('Cannot obtain remote version: Not connected to internet')
         return None
 
-    try:
-        log.debug('Obtaining remote version from github')
-        get_remote_tags()
-        references = check_output(
-            ['git', 'ls-remote', '-t', '-q', '--exit-code', '--refs', 'origin', 'tags', '*'],
-            cwd=BASE_DIR,
-            timeout=5,
-            encoding='UTF-8',
-        )
-
-    except (SubprocessError, CalledProcessError, FileNotFoundError):
-        log.error('Could not obtain remote version string')
-        return None
+    references = call_git('ls-remote', '-t', '-q', '--exit-code', '--refs', 'origin', 'tags', '*', on_error='log')
 
     try:
         log.debug('Parsing local version string')